Move timestamp/size formatting helpers into core
format_systemd_timestamp, format_duration, and format_bytes were living in
the TUI's home component but have nothing TUI-specific about them. The GUI
needs them for timer trigger times, so they now live in
systemctl_ui_core::format. No functional changes.
