+++ crates/systemctl-tui/src/components/home.rs
@@ -26,6 +26,7 @@ use std::{collections::HashSet, process::Stdio, time::Duration};
 use super::{logger::Logger, Component, Frame};
 use crate::{
   action::Action,
+  format::{format_bytes, format_duration, format_systemd_timestamp},
   journal::{parse_json_log_line, LogEntry},
   systemd::{
     self, diagnose_missing_logs, parse_journalctl_error, Scope, UnitFile, UnitId, UnitKind, UnitRuntimeInfo, UnitScope,
@@ -2538,62 +2539,6 @@ impl Component for Home {
   }
 }
 
-/// Parse a journalctl timestamp and return a formatted date string.
-///
-/// systemd v255 changed the timestamp format from `-0700` to `-07:00` (RFC 3339).
-/// See: https://github.com/systemd/systemd/pull/29134
-/// Parses a systemd "show" timestamp like "Wed 2026-07-08 10:00:00 PDT" into a compact
-/// absolute form and, for local units, a relative one ("2d 4h ago" or "in 2d 4h").
-/// Remote timestamps retain their source timezone and omit the relative time because
-/// comparing naive wall-clock values from different timezones would be misleading.
-fn format_systemd_timestamp(timestamp: &str, is_remote: bool) -> Option<(String, Option<String>)> {
-  let mut parts = timestamp.split_whitespace();
-  let _weekday = parts.next()?;
-  let date = parts.next()?;
-  let time = parts.next()?;
-  let timezone = parts.next();
-  let naive = chrono::NaiveDateTime::parse_from_str(&format!("{date} {time}"), "%Y-%m-%d %H:%M:%S").ok()?;
-  let mut absolute = naive.format("%Y-%m-%d %H:%M").to_string();
-  if is_remote {
-    if let Some(timezone) = timezone {
-      absolute.push(' ');
-      absolute.push_str(timezone);
-    }
-    return Some((absolute, None));
-  }
-  let seconds = chrono::Local::now().naive_local().signed_duration_since(naive).num_seconds();
-  let relative = if seconds >= 0 {
-    format!("{} ago", format_duration(seconds as u64))
-  } else {
-    format!("in {}", format_duration(seconds.unsigned_abs()))
-  };
-  Some((absolute, Some(relative)))
-}
-
-fn format_duration(seconds: u64) -> String {
-  match seconds {
-    s if s < 60 => format!("{s}s"),
-    s if s < 3600 => format!("{}m {}s", s / 60, s % 60),
-    s if s < 86400 => format!("{}h {}m", s / 3600, (s % 3600) / 60),
-    s => format!("{}d {}h", s / 86400, (s % 86400) / 3600),
-  }
-}
-
-fn format_bytes(bytes: u64) -> String {
-  const UNITS: [&str; 5] = ["B", "K", "M", "G", "T"];
-  let mut value = bytes as f64;
-  let mut unit = 0;
-  while value >= 1024.0 && unit < UNITS.len() - 1 {
-    value /= 1024.0;
-    unit += 1;
-  }
-  if unit == 0 {
-    format!("{bytes}B")
-  } else {
-    format!("{value:.1}{}", UNITS[unit])
-  }
-}
-
 /// Greedy word-wrap to a given column width. Returns one string per visual line. Words longer than
 /// the width are placed on their own line and allowed to overflow (rare for prose). Width is
 /// measured in chars, which matches display width for the plain English these descriptions use.
@@ -2887,13 +2832,6 @@ mod tests {
 
     assert!(home.menu_items.items.iter().any(|item| item.name == "Start backup.service now"));
     assert_eq!(home.menu_items.selected().map(|item| item.name.as_str()), Some("Disable timer"));
-  }
-
-  #[test]
-  fn remote_timestamps_keep_their_timezone_without_a_relative_guess() {
-    let formatted = format_systemd_timestamp("Wed 2026-07-08 10:00:00 PDT", true);
-
-    assert_eq!(formatted, Some(("2026-07-08 10:00 PDT".into(), None)));
   }
 
   #[test]
