+++ crates/systemctl-tui/src/lib.rs
@@ -14,4 +14,4 @@ pub mod remote_picker;
 
 // Re-exported from the shared core crate so existing `crate::systemd`-style
 // paths keep working in both this crate and downstream code.
-pub use systemctl_ui_core::{journal, ssh, systemd, unit_descriptions};
+pub use systemctl_ui_core::{format, journal, ssh, systemd, unit_descriptions};
