+++ crates/systemctl-ui-core/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod format;
 pub mod journal;
 pub mod ssh;
 pub mod systemd;
